requirements.txt
# MAKCU AIO Requirements

# GUI Framework
customtkinter>=5.0.0

# Serial Communication
pyserial>=3.5

# HTTP Requests (for updates and community profiles)
requests>=2.28.0

# Image Processing
Pillow>=9.0.0

# Retry Logic
tenacity>=8.0.0

# USB Communication (optional, for advanced features)
pyusb>=1.2.0

# For building executable (development only)
pyinstaller>=5.0
